functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin
firebase-functions
Pillow
qrcode
google-cloud-storage